requirements.txt
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.46
iniconfig==2.3.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.2
python-dateutil==2.9.0.post0
rich==14.3.3
six==1.17.0
smmap==5.0.3
tzdata==2026.1
